pyproject.toml
[tool.poetry]
name = "cloud_sheets_slim"
version = "0.2.1a0"
homepage = "https://github.com/yingca1/cloud-sheets-slim"
description = "A lightweight cloud sheets operation library"
authors = ["Ying Cai <i@caiying.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
gspread = "^5.12.0"
gspread-pandas= "v3.2.2"
pandas = "^2.1.4"
oauth2client = "^4.1.3"
python-dotenv = "^1.0.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
